requirements.txt
asgiref==3.9.1
beautifulsoup4==4.14.3
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
Django==4.2.23
django-bootstrap3==25.3
django-bootstrap4==25.3
django-filter==25.2
django-secret-key==1.0.1
django-taggit==6.1.0
environs==14.5.0
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
marshmallow==4.1.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.8.1
sqlparse==0.5.3
stripe==12.4.0
telegram==0.0.1
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
